pyproject.toml
[tool.poetry]
name = "gtm-mcp"
version = "0.1.3"
description = "MCP server for Google Tag Manager"
authors = ["Paolo Bietolini"]
license = "MIT"
readme = "README.md"
packages = [{ include = "gtm_mcp", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.186.0"
google-auth-oauthlib = "^1.2.3"
mcp = "^1.20.0"
protobuf = "^6.33.0"
python-dotenv = "^1.2.1"

[tool.poetry.scripts]
gtm-mcp = "gtm_mcp.server:run"
gtm-mcp-setup = "gtm_mcp.setup:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
